Fully support `DECRQM` (Request Mode Setting).
This causes a `DECRPM` (Report Mode Setting) control sequence to be
written back to the terminal input, with the mode's setting. This allows
the current state of any mode set with an `SM`/`RM` or `DECSET`/`DECRST`
(private `SM`/`RM`) sequence to be queried.

This expands `DECRQM` support to all modes supported by tmux.

Also, add private mode 14. In xterm, this mode indicates whether
`ATTCUBL` and mode 13 are XORed or inclusive-ORed. Report it as
"inclusive-OR" so applications can interpret modes 12 and 13 properly.