Fully support `DECRQSS` (Request Selection or Setting).
This is a DCS that is used to query a terminal setting that was set by
certain control sequences. It causes a `DECRPSS` (Report Selection or
Setting) DCS containing the sequence to set the current setting, minus
the CSI bytes, to be written back to the terminal input.

This was already supported for the cursor type (`DECSCUSR`); this change
expands it to support `SGR` and `DECSTBM` also.

This also fixes tmux not to repeat the input string in the `DECRPSS`
string. DEC STD 070 states on page 5-197:

> The data returned by the DECRPSS control string is the entire control
> function that the host would send to the terminal to select or set the
> parameter being inquired, based on the current setting or selection of
> that parameter...
>
> For example, if the current Graphic Rendition is blinking, and
> reversed, the data would consist of the characters '0;5;7m', which is
> the parameter string and final character necessary to change the
> graphic rendition to that state.

Existing terminal emulators such as xterm do not repeat the input string
in the `DECRPSS` string.