Support `DECRSPS` (Restore Presentation State).
This DCS restores a setting that was previously reported with
`DECRQPSR`. It differs from `DECPSR` only in the final command byte.

I'm not entirely happy with the `DECCIR` implementation. Only the four
`SGR` attributes that were supported up to the VT420 (bold, underscore,
blink, reverse) are restored. The problem is that, although it contains
provisions for extension (bit 4 of each flags char), DEC never extended
`DECCIR` AFAIK to support the invisible and color attributes that were
supported by the VT525. I suppose it's theoretically possible to extend
this ourselves; the problem is getting the community to agree.