Support setting the level of VT emulation.
The initial level can be set with the new `default-emulation-level`
option. This affects the response to a `DA` sequence, as well as the
available private functions. Available private functions can also be
constrained with the `DECSCL` (Set Conformance Level) sequence.

Correct the `DA` response when SIXEL is supported. The '4' parameter is
of course not supported unless the first parameter in the response is in
the range [61,69]. The VT100 response means something different entirely
("I am a VT100 ('1') with the Advanced Video Option, but not the STP or
graphics options ('2').")