Add a new "exit-time" option.
This sets the amount of time that `tmux(1)` will wait when killing a
pane for the process tree to clean itself up, before it proceeds to tear
down the pane. This allows, e.g. saving the contents of the pane in an
exit hook, so they can be restored when `tmux(1)` starts the pane again.