Use `OSC` 1337 to query terminal features.
The proposal is [here][1].

AFAIK, no terminal implements `XDA` 1, and only iTerm implements `OSC`
1337, which isn't even part of the current draft of the spec. Also, one
of the flag strings is ambiguous, which makes parsing somewhat _ad hoc_
and fragile. Further, this spec is still a proposal, and it includes a
proposal to use a base64-encoded binary string instead of the text
string that iTerm currently emits; telling the two apart could pose a
problem.

[1]: https://gitlab.freedesktop.org/terminal-wg/specifications/-/blob/45c1c6c46e0de498064e826f63e1daf3ab400f82/proposals/feature-reporting/index.md