Back up twice when the cursor is one beyond the right margin.
The cursor would be about to wrap in that case; but on a real VT100 in
this situation, the cursor would be exactly _at_ the right margin with
the wrap flag set. Therefore, a program designed for the VT100 would
expect the cursor after a backspace to be one _before_ the margin, and
not wrap.

Fixes the `DECAWM` test from `vttest`.